remove anonymous namespace from make_enum
