+++ src/software/util/make_enum/BUILD
@@ -2,6 +2,7 @@ package(default_visibility = ["//visibility:public"])
 
 cc_library(
     name = "make_enum",
+    srcs = ["make_enum.cpp"],
     hdrs = ["make_enum.h"],
 )
